pkg/destroy/gcp: stop instances before deleting resources
Previously, we were in a race condition where the cloud controller might
re-create resources that we have already deleted. Some of these
resources were not discoverable because the discoverable resources (such
as target pools) were not yet recreated. This change stops all instances
before delting things in order to avoid the race alltogether.