Revert "Merge pull request #2471 from openshift-cherrypick-robot/cherry-pick-2425-to-release-4.2"
This reverts commit 8a194c3c24c52d17455d525a69282400e5067cbe, reversing
changes made to e349157f325dba2d06666987603da39965be5319.

We found in BZ https://bugzilla.redhat.com/show_bug.cgi?id=1762618 that at least
some of our code, running on cluster, was accessing the external instead of internal
api.

While it is possible to make it work (by allowing your proxy to talk from internal
to the apiserver) this is a change/regression from GA. Since we hit this with at
least 1 part of our code we have decided to revert and try again with more testing.