OpenStack: unpack compressed image data
Now we always download unpacked image data to the local filesystem
and then upload it to Glance without any modifications.

In the future we plan to provide only compressed data, but
unfortunately OpenStack Glance is not able to work with it. So, we
have to decompress the data first.

This patch allows the installer to work with both compressed and
uncompressed images.
When compressed image is downloaded the installer decompresses the
data into local cache and returns its path. For compressed images
the workflow stays the same.