pkg/asset/ignition: bootstrap kubeconfig to use api-int
Prior to this change, the default kubeconfig on the bootstrap host used
the external api url. This caused issues when the external api was not
available such as when proxy is configured but the noProxy list does not
contain the external api. In such situation, the bootstrap host fails to
report bootstrap complete to the cluster. As a result, the installer
fails during the wait-for bootstrap-complete stage.

This change adds a new kubeconfig (adminInternal) which points to
api-int. The bootstrap host defaults to this new kubeconfig. Api-int
is already in the noProxy list. As a result, the bootstrap host can
report complete even when the external api is inaccessible.