libvirt: Resize specifically to 16G
Today for IaaS clouds, we default to an instance type, which
then in turn usually provides a default size.  RHCOS resizes on
boot to that size, distinct from its "default" 16G size.

However, libvirt installs were inheriting our default size.  We'd
like to shrink it because we plan to land encryption:
https://github.com/openshift/enhancements/pull/15
And the less data we need to encrypt, the better.

(In the future I'd like to make this configurable with a variable,
 but let's just prepare for the encryption work now)