pkg/asset/installconfig/aws: Serialize region and subnets
Because when the asset store loads the InstallConfig asset from state,
it wasn't initializing the Metadata structure (vs. generating a new
InstallConfig or loading it from a file, both of which hit
InstallConfig.finish() and its NewMetadata call).  With this commit,
we store copies of the region and subnets (if set) in the serialized
Metadata state to avoid [1]:

  $ openshift-install create ignition-configs
  $ openshift-install create cluster
  FATAL failed to fetch Terraform Variables: failed to generate asset "Terraform Variables": no subnets configured

I'm not super-excited about having multiple copies of this data in the
state file, but the alternative would be to add Finish() to the Asset
or WritableAsset interfaces, and that seems a bit heavy too.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1771431