[release-4.2] Bug 1771358: manually prepare cloud provider's config
We can't use "go-ini" library to generate the config, because it's
incompatible with "gcfg" library that we use to read the data.

For instance, if there is a string with a # character, then "go-ini"
wraps it in bacticks, like `aaa#bbb`, but gcfg doesn't recognize it
and  parses the data as `aaa, skipping everything after the #.

So, we have to generate the config manually, avoiding possible issues
with "go-ini".