ovirt: BaseImage handling
oVirt needs to upload the image which is pinned to the version
of the installer, and create an ovirt template from it, specific to the
installer openshift cluster.

Since oVirt uses the same openstack image(for the ignition via
cloud-init disk suport), I exposed openstack handling of the image and
made both platform to invoke it.

The result is that the image info is extracted from the version, the
image is downloaded if not already under the cache, then tf uploaded an
image to oVirt and creates a template out of it, and the template ID is
used in the install config.

Depends-on: https://github.com/oVirt/terraform-provider-ovirt/pull/184

Signed-off-by: Roy Golan <rgolan@redhat.com>