upi/metal: Set the facility to SJC1
I was seeing masters within a cluster being splayed across SJC1 and EWR1
datacenters and that can contribute to etcd latency which may lead to
some of the E2E failures we're seeing.

Future improvement could be to provision the bootstrap host, determine
its facility and then ensure all the remaining hosts are provisioned
into the same facility.