data/gcp: use separate instance-group for bootstrap-host
k8s will use any instance group that only has instances that are `nodes` of the cluster when creating internal load balancer.
Since previously, the bootstrap-host, which is not a k8s node for the cluster, is added to the instance-group for one of the control-plane, k8s skips that instance-group and tried to add that control-plane node to the internal load balancer using one of it's own instance-group.
The side effect of the actions is that particular instance-group is marked for the subnet of the control-plane and then adding workers from other subnet keeps failing after that.

```
Validation failed for instance 'projects/openshift-dev-installer/zones/us-east1-b/instances/padill-swswf-m-0': instance may belong to at most one load-balanced instance group.
The kube-controller-manager logs may contain more details.
```

This creates a separate isntance-group for the bootstrap-host