Bug 1776272: OpenStack: fix getting service catalog
We need to obtain the service catalog in several places inside the code,
but each time we expect that project_name is specified in the clouds.yaml,
and if it's not there, installation fails.
In general it's not necessary and users can specify just project_id instead
of the name.

This commit updates the service catalog downloading functions by removing
this restriction.