Simplify bootstrap ignition config generation
Now, when we have a generated ignition config, additionally we inject
invoker information into it during TF vars generation.
This approach has downsides. First, we inject this data into the
marshalled json file and not in the config structure itself, which
means they are not the same after the change. Second, modification of
the asset in several places in the code leads to fuzzy architecture and
potential errors.

This commit moves the data injection from the tfvars module to the
bootstrap ignition module, where the asset is origionally generated.