OpenStack: POC: Swiftless bootstrap
This PR allows to remove the Swift storage requirement for creating
bootstrap machines.

Now we use Swift to store ignition configs because they are too big
for Nova's metadata service (300kb vs 16kb available). This commit
allows to shrink the configs by extracting data that is not required on
the initial steps of bootstrapping. Later we upload the remaining files
to the machine using the sftp client.