baremetal: make provisioning networks more configurable
This change adds 4 new options to the baremetal platform:

        - ProvisioningNetworkInterface
        - ProvisioningNetworkCIDR
        - ProvisioningDHCPExternal
        - ProvisioningDHCPRange

This makes the provisioning network more configurable, by allowing users
to specify a specific network interface, which network, and settings for
whether the cluster should provide DHCP services on the provisioning
network, or if it will be managed externally. If we will provide DHCP
services on the network, the user may further customize which range to
use.

This change is a prerequisite to a number of additional high-priority
work items for the baremetal IPI platform, including single-stack IPv6
provisioning, external DHCP support, and end-to-end seamless deployment
without a manually created config map.

Co-authored-by: Ian Main <imain@redhat.com>
Co-authored-by: Stephen Benjamin <stephen@redhat.com>