Bug 1777232: Add GCP-specific validation of cluster name.
Validation of the Cluster Name in the wizard or metadata.name in the install-config only ensures validity as a domain name. The cluster name is also used to name resources created by the installer, and this commit adds validation to ensure that the domain name meets the requirements for naming resources in GCP. This has caused failed installs in GCP when cluster names begin with numbers.