rhcos: Bump to 43.81.201912110942.0
This has encryption/FIPS fixes; in particular, we now no
longer require a TPM2 by default on bare metal.

Note that while RHCOS has a 4.4 build, it's really almost the
same as 4.3 right now, so I simplified things and just used the
actual 4.3 builds to match the 4.3 branch.

We'll swap to RHCOS 4.4 when it's really 4.4.