terraform/plugins: update libvirt provider
This pulls in some upstream changes [1] that add support for providing
Ignition configs to s390x guests. This doesn't require any changes to
the Terraform templates, but it does require Ignition to be run with
`--provider=openstack` on s390x (there is no support for the QEMU
firmware configuration device on s390x, so the OpenStack config-drive is
used instead).

[1]: https://github.com/dmacvicar/terraform-provider-libvirt/pull/678