Bug 1780534: create bootstrap floating ip after the bootstrap machine
Now we create the bootstrap floating ip in the beginning of the
installation and then wait the bootstrap machine to become ready.
If the provisioning takes more than 10 minutes, due to long image
uploading or other reasons, fip creation fails with timeout.
To prevent this we should create the fip when the machine is already
available.