WIP: Never disable hyperthreading and add tsx=on to kernel params
Enable Transactional Synchronization Extensions [1].  These were
disabled by default for CVE-2019-11135 [2,3], but we think the change
is causing a perfomance impact.  Re-enable it to see if we recover.

While testing this, I'm stealing some scaffolding which was previously
used for disabling hyperthreading.  We want both TSX and
hyperthreading for this test.  If it turns out to be useful, we can
work out if/how we want an install-config TSX knob to work.

[1]: https://en.wikipedia.org/wiki/Transactional_Synchronization_Extensions
[2]: https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2019-11135
[3]: https://access.redhat.com/articles/tsx-asynchronousabort