cmd/openshift-install/gather: Gather bootstrap console logs
If we can't reach the bootstrap machine via SSH.  Before this commit,
we would occasionally see connection issues like [1]:

  level=info msg="Waiting up to 30m0s for the Kubernetes API at https://api.ci-op-6266tp8r-77109.origin-ci-int-aws.dev.rhcloud.com:6443..."
  level=error msg="Attempted to gather ClusterOperator status after installation failure: listing ClusterOperator objects: Get https://api.ci-op-6266tp8r-77109.origin-ci-int-aws.dev.rhcloud.com:6443/apis/config.openshift.io/v1/clusteroperators: dial tcp 3.221.214.197:6443: connect: connection refused"
  level=info msg="Pulling debug logs from the bootstrap machine"
  level=error msg="Attempted to gather debug logs after installation failure: failed to create SSH client, ensure the proper ssh key is in your keyring or specify with --key: dial tcp 3.84.188.207:22: connect: connection refused"
  level=fatal msg="Bootstrap failed to complete: waiting for Kubernetes API: context deadline exceeded"

With this commit, when we see those connection-refused errors, we
attempt to retrieve console logs for the bootstrap instance.  This
will make it easier for users and users to see why the machine failed
to come up.  It should be especially useful in continuous integration
when bumping RHCOS boot images [2], when such boot-time failures are
more likely.

I've only implemented it on AWS for the moment, but I've set it up so
we can extend it to other platforms going forward.

[1]: https://prow.svc.ci.openshift.org/view/gcs/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-upgrade/12076
[2]: https://github.com/openshift/installer/pull/2777#issuecomment-565237752