vendor: Switch to a branch of azurerm that fixes three bugs
The first is that private_ip_address_version is not supported on
load balancers in 2018-12-01 azure, but is in 2019-04-01. Specifically
override that one call until we bump.

The second is that the ip_version for static allocation check on the
load balancer is out of date for Azure - LBs can have both types of
allocation in the Standard SKU since Azure added preview of the type.
Remove the check like we would for an upstream patch.

The third is to add support for subnets to have multiple address
prefixes which is required to add dual stack support. This is an
upstream change.