azure: Add full dual-stack IPv6 support to the install for Azure
The installer can now provision a dual-stack IP v4/v6 (v4 primary)
cluster if OPENSHIFT_INSTALL_AZURE_USE_IPV6=true is passed on
install. Unlike the AWS configuration, this is currently configured
to make IPv6 transparent. The machine CIDR for ipv6 is not
configurable and is currently hardcoded to fd00::/48.

A full dual stack configuration would in theory look like the
following (barring the missing machineCIDR configurability).

```
networking:
  clusterNetwork:
  - cidr: 10.128.0.0/14
    hostPrefix: 23
  - cidr: fd01::/48
    hostPrefix: 64
  machineCIDR: 10.0.0.0/16
  networkType: OVNKubernetes
  serviceNetwork:
  - 10.1.0.0/16
  - fd02::/112
```

This commit will be updated to include the ipv6 primary mode (if
all input addresses are ipv6) and to not require the use of the
environment variable.