types: Validate networking options for single or dual stack
Add validation that checks the networking section of install-config.yaml
for consistent with regards to IPv6 and dual-stack support.

Rules:

* If any networking list (service, pod, or machine) has IPv6 and IPv4
  CIDRs, all fields must have both sets of addresses
  * Except on AWS, which allows the machine field to only have a single IPv4 CIDR
* Only BareMetal and None can provide dual-stack or ipv6 addresses
  * Will be changed by forthcoming PRs that add support for AWS and Azure
  * Other platforms may have a toleration mode for UPI that will be relaxed in 4.4
* Block OpenShiftSDN when IPv6 is requested since it is not implemented
* ServiceNetwork must always have 1 or 2 entries, for single or dual stack

These rules may be altered by subsequent changes if we need to relax requirements.

A helper method is provided to determine which mode is requested for a given
install-config which should be used during asset generation per platform.

Relax validation on SubnetCIDR and move logic to this new function.