openstack UPI: Do not rely on auto_ip to create bootstrap FIP
The `auto_ip` option to `os_server` is unreliable as it may not pick an
IP address from the floating IP pool we want in case the cloud has
multiple floating IP pools. Instead we create the FIP and assign it to
the node in a separate task.

We need to explicitely disable `auto_ip` as the default value is yes.