ovirt: optionally use a different logical network for instances
Motivation
Some installatino may want to use different network to segregate the
installation VMs from other existing workloads and/or to make use of
specific network which are not 'management' network, that has traffic
dedicated to the workload.

Change
installconfig has an optional ovirt_network_name and should be an
existing network name for the selected ovirt_cluster_id.
The default value is 'ovirtmgmt' which is the default network for any
ovirt cluster.

Note: because of lack of support from the terraform provider I had to
use the network name and not ID.

Signed-off-by: Roy Golan <rgolan@redhat.com>