Ensure enough number of retries for cluster with Kuryr SDN
When using Kuryr SDN the number of resources created on the
OpenStack side is way larger than for the OpenShift SDN case.
For this reason, as there is no specific order for the resources
to be deleted, the chances of the deletion of one resource (e.g.,
a network) being blocked by the deletion of another (e.g. a port
on that network) are higher, thus needed more retries.