aws: Explicitly flag on ena_support
We want enhanced networking in AWS.  Now, it turns out that today
we have this in fact because `ore` in the latest
https://github.com/coreos/coreos-assembler turns this on
explicitly.  And when Terraform goes to copy AMI with encryption,
it preserves the value of that variable.

But, in the future we want to drop the AMIs from the RHCOS
pipeline and just upload a blob (like Azure/GCP), so let's add
this reminder to ourselves that we need to enable ENA when we
do that.

xref https://github.com/openshift/installer/issues/2906