pkg/terraform/exec/plugins: Switch to Azure 1.27.1 fork with Azure fixes
The azurerm provider does not fully support IPv6. Until that time, switch
to a fork of the provider based on 1.27.1 plus three specific fixes that
allow dual stack configuration on Azure.

When we switch to newer azurerm/terraform, we can drop these.

In order to get load balancers to set PrivateIPAddressVersion, switch
to a forked version of azure-sdk-for-go from v26.7.0 that allows that
field to be set only on load balancers.