asset/manifests: add openshift-install configmap
This adds support for generating the openshift-install ConfigMap when
using a UPI installation flow. Before this, the ConfigMap was generated
during the "cluster" target rather than the manifests target, so that we
could distinguish between an IPI and a UPI installation. The user had
the option of overriding the "invoker" field using the environment
variable OPENSHIFT_INSTALL_INVOKER. Even if the environment variable was
used, the ConfigMap would not be present in a UPI installation (because
it wasn't generated in the "manifests" target). This change causes the
installer to generate that ConfigMap in the "manifests" target (in
addition to the "cluster" target) when it sees the environment variable.

We will make use of this functionality in CI, where we specify the job
as the invoker.