cmd/openshift-install/create: add env var to preserve bootstrap
For a variety of reasons, it may be useful to keep the bootstrap VM
around for debugging. My general approach has been to interrupt the
installer before it has a chance to delete the bootstrap VM.

If the environment variable OPENSHIFT_INSTALL_PRESERVE_BOOTSTRAP is set
to any value, the installer will not delete the bootstrap.