Set default CPUs and Memory for vSphere nodes.
Sets the vCPU and memory to the minimum requirements for both the control plane and worker nodes. Includes the default values in the machinesets as well as terraform code for creating the control plane.

This commit also includes documentation for customizing the machine pools. Note that the memory json specification in the install-config uses MB while the variable name is MiB. The intention of this split is to keep the user-facing docs and install-config consistent with Red Hat docs which use MB/GB, but the variable names to be consistent with the machine-api and kubernetes in general which uses MiB/GiB.