script: Set GCP validation to precede domain validation.
When the user chooses to create an install config using the installer console, picks the GCP environment and gives an improper cluster name, the error message that is being displayed now is that the name fails to match the norms of openshift but a better error message is to show that GCP needs the name in a specific format.

Another bug was found when the user loads the install configs with an improper cluster name, both the GCP error message and the openshift error messages were being displayed. Fixed the UI errors.