script: Added Azure validation for cluster names
The Azure platform requires a specific format for the names of the clusters which was not present in the installer. Added the checks during both the creation and loading of the install config files to check for Azure validation specifically before trying to create the cluster on Azure if the user chooses said cloud platform.