create.go: add the logic of waiting for the etcd cluster to go
In the GCP environment, the bootstrap node cannot communicate with
the kube api in order to make sure that the etcd cluster is up before
marking bootstrap complete. This is because, it needs to stop the
gcp-routes service after bootkube.sh exits.

Thus, this patch removes the logic of waiting for the etcd cluster from
bootkube.sh and adds it to cmd/openshift-install/create.go