RHCOS: revert to 44.81.202001241431.0
In https://github.com/openshift/installer/pull/2986 we updated RHCOS
to 44.81.202001241932.0, which happened to carry a COSA change
https://github.com/coreos/coreos-assembler/commit/1dc3cddb393fcab1000ee62908aaadb8d41d8ccd
which broke GCP metadata. Since we didn't run e2e-gcp on that 2986,
we didn't catch this. This seems to have broken GCP CI since it is
unable to find the tarball being referenced in the meta.

This temporarily reverts RHCOS to one version previous, which includes
the IPV6 changes but doesn't have the GCP break. This should fix GCP
for now as we fix COSA.

In the future we should probably have each RHCOS bump go through
all the platforms.

Signed-off-by: Yu Qi Zhang <jerzhang@redhat.com>