data/bootstrap: delay the removal of bootstrap mcs
The bootstrap MCS is responsible for serving Ignition configs to the
booting control plane machines while we wait for the full control plane
(which will host the in-cluster MCS). If we remove this too early, it
will prevent the control plane machines from booting and the control
plane from starting. The etcd health check, which is the gate before we
remove the MCS, reports healthly (since the one-node, bootstrap etcd
cluster is actually up) before the control plane machines had a chance
to boot. The result is that one or more of the control plane machines
get stuck during boot (because Ignition is still trying to fetch a
config) and the cluster fails to bootstrap.

This moves the bootstrap MCS removal after bootkube has finished, which
will ensure that all of the initial manifests have been loaded into the
cluster (including the MCO/MCS manifests).