VSphere: Strip :<port> from vCenter in cloud-creds-secret
VCenter requires that the cloud-creds-secret contain the workstation in the key of the username and password. e,g:

data:
   10.0.1.200.username: QWRtaW5pc3RyYXRvckB2c3BoZXJlLmxvY2Fs
   10.0.1.200.password: cGFzc3dvcmQ=

This is for compatibility with upstream storage: https://vmware.github.io/vsphere-storage-for-kubernetes/documentation/k8s-secret.html

We need to strip any colons and ports that are included in the workstation, as they are illegal keys in secrets file.