aws: pick instance types based on selected availability zones
Using new AWS API DescribeInstanceTypeOfferings [1] we can now check what instance types are available for specific region/AZs, which allows us to check which instance types are available in the selected AZs.
Since we default to m4 for most regions and only override to m5 for certain, in cases where new AZs are added to older regions like sa-east-1, see BZ 1774465 [2], we end up picking m4 when m5 is present in all the chosen AZs.

Now we can define a list in descreasing priority order for instance classes for a region and the installer can try to pick default based on what works for all the selected AZs. We still do not support heterogenous setup which different instance types in different regions, but this
solves the case when newer m5 needs to be used in certain regions based on the AZ selection.

The default order for instance classes stays m4 > m5.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_DescribeInstanceTypeOfferings.html
[2]: https://bugzilla.redhat.com/show_bug.cgi?id=1774465