destroy/gcp: also remove instances with owned label.
Prior to this change, the gcp destroy process would get stuck in an
infinite loop in the case where an instance has been provisioned in
a cluster subnetwork but the instance name does not contain the infra_id
identifier. As a result, the instance would be ignored and the google
api would complain that the subnetwork was still in use when attempting
to delete it. This can be experienced after adding new machine groups to
a cluster such that the new instance names do not contain the infra_id.

This change enables the destroy process to also delete instances that are
labeled with kubernetes_io_cluster_infra_id:owned. This ensures all of
the instances created by the cluster are deleted. As a result, cluster
provisioned instances will no longer block on deletion of subnetworks.