Bug 1796347: Move permissions check to cluster asset
Separates the permissions check logic from PlatformCredsCheck and creates a similar PlatformPermsCheck. While PlatformCredsCheck checks whether credentials can authenticate to a platform, the new PlatformPermsCheck ensures that an account has the necessary permissions to create a cluster. PlatformPermsCheck is only needed when a cluster is being created and is currently only implemented for AWS.

This fixes BZ-1796347 where UPI calls to create manifest were failing due to insufficient permissions.