AWS: Allow users to pick the Zone ID when there are multiple DNS Zones with the same name
When there are multiple zones with the same name, we give the user the option
of choosing the zone ID. For interactive installs, the zone ID will display
with the zone when there are duplicates. For non-interactive, the zone ID is
specified in the install-config file. The zone ID is then pushed down to the
terraform where the correct zone is used.

- https://jira.coreos.com/browse/CORS-1070
- https://bugzilla.redhat.com/show_bug.cgi?id=1733389