Baremetal: Be explicit that addresses must be in same subnet
When starting the static pods on a host, we determine the correct
address to use by comparing its subnet with the VIP subnet. If the
external address assigned by DHCP is in a different subnet from the
VIPs then we can't determine which address to use and deployment
will fail. This change makes the requirement for addresses to be in
the same subnet explicit.