Add new BootMode field
Create a new field in BareMetalHost to indicate
boot mode. It can have UEFI and legacy values (defaulting
to UEFI). The boot_mode capability will be generated using
that field instead of hardcoding it. Also the right value
will be written in BareMetalHost definition, to pass the
right value to workers.

Signed-off-by: Yolanda Robla <yroblamo@redhat.com>