Bug 1809780: OpenStack: pass all CAs to bootstrap ignition
When the CA cert file referenced from the `clouds.yaml` contained more
than one certficate, the complete cert bundle was added to the boostrap
ignition file, but only the first one was actually trusted to gather
the ignition file.

This patch splits the CA cert bundle into separate certs and add them
to the ignition file as separate entries.