validation: Allow periods in gcp and azure cluster names.
Prior to this change, GCP and Azure did not allow periods in the cluster
names because that would violate their API naming conventions. However,
newer versions of the installer create cloud assets based on the InfraID,
where all periods are now translated to dashes.

This change modifies the verification to allow periods. As a result,
users can now create clusters with names such as "my.prod". They must
otherwise comply with RFC1035 "[a-z]([-a-z0-9]*[a-z0-9])?".