aws: AWS China support
Prior to this change, installer will failed to run on AWS China regions,
because route53 currently is not GA in China, the api endpoint is
not in aws sdk.

With this change, if we create an install-config, with custom AMI ID
set, and China region set, we can successfully install a cluster on AWS
China.