OpenStack: Fail early when boostrap ignition shim is to big
Nova has a hard limit of 64KB [1] for user data when booting VMs. This
was not a problem previously when everything that went into the user
data was under our control and small enough. With the recent addition
of self-signed certificate feature, the ignition size now depends on
the size of the CA cert file the user passes, and it may cause it to
exceed nova limit if the file is too large.

This commit checks the size of the rendered file as it would be passed
to nova user data and fails early with an appropriate message if it
exceeds the 64KB limit.

[1] https://docs.openstack.org/nova/latest/user/metadata.html#user-data

Fixes #3177