libvirt: Use OpenStack rhcos image instead of Qemu for s390x and ppc64le
The Qemu image on non x86 architectures (specifically s390x, ppc64le) suffers from a problem
that it does not support fw_cfg which is used to inject ignition config. As of today there is
no alternative to that in Qemu. A while ago config drive support was added to s390x and ppc64le
for the libvirt environment to get around that problem. But currently that fix is usable only with
the OpenStack image. So, with this fix , the OpenStack image is picked automatically for a libvirt
install until we have something similar to fw_cfg for s390x and ppc64le.