data: Tune all cloud load balancers to have consistent intervals
In general we want shorter health check intervals to detect apiserver
failures faster. The current defaults are on the long side, and we
have already standardized ingress to use the minimum intervals possible.

Azure and AWS can be tuned to 5s intervals, so make them consistent.
GCP is more efficient and can be tuned to 1s, but that may increase
idle load on the API servers, so set the default to 2s with 3 checks
to keep it closer to Azure and AWS.