openstack UPI: Adapt the script to older Ansible
The Ansible version that ships with RHEL does not have knowledge of Nova
microversions from 2.15; they will block the 'soft-anti-affinity'
policy.

This change replaces the use of the os_server_group module with a raw
HTTP call to the Nova API, in order to specify microversion 2.15, the
first compatible with soft-anti-affinity.